requirements.txt
Django==5.1.6
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
celery==5.4.0
redis==5.2.1
openpyxl==3.1.5
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
coverage==7.6.10

xlrd==2.0.1
